Moved CSS files to their own folder
