+++ front-end/src/containers/Home.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './Home.css';
+import '../styles/Home.css';
 import { Link } from 'react-router-dom';
 
 class Home extends Component {
