+++ front-end/src/containers/Navbar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../App.css';
+import '../styles/App.css';
 
 function Navbar() {
     return (
