+++ front-end/src/containers/NotFound.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./NotFound.css";
+import "../styles/NotFound.css";
 
 export default function NotFound() {
   return (
