+++ front-end/src/styles/Home.css
